feat: rename structure
